Split out SchemaGenerator
